Refactor : Cgange directory_name view into controller in repo.
